refactor services

+++ SimpleTrader/SimpleTrader.WPF/Features/Accounts/Models/Account.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using FieldOps.Kernel.Entities;
-using SimpleTrader.WPF.Features.Assets;
 using SimpleTrader.WPF.Features.Assets.Models;
 using SimpleTrader.WPF.Features.Users.Models;
 
@@ -9,7 +8,7 @@ namespace SimpleTrader.WPF.Features.Accounts.Models;
 
 public class Account : EntityBase
 {
-    public double Balance { get; set; }
+    public decimal Balance { get; set; }
     public Guid AccountHolderId { get; set; }
     public virtual User? AccountHolder { get; set; }
     public virtual ICollection<AssetTransaction> AssetTransactions { get; set; } = [];
